Refactored helper methods into the Criterion class
