+++ app/controllers/criteria_controller.rb
@@ -1,5 +1,4 @@
 class CriteriaController < ApplicationController
-  include CriteriaHelper
   before_action { authorize! }
 
   layout 'assignment_content'
@@ -158,7 +157,7 @@ class CriteriaController < ApplicationController
     else
       if data[:type] == '.yml'
         ApplicationRecord.transaction do
-          successes = upload_criteria_from_yaml(assignment, data[:contents])
+          successes = Criterion.upload_criteria_from_yaml(assignment, data[:contents])
           flash_message(:success, I18n.t('upload_success', count: successes)) if successes > 0
         rescue StandardError => e
           flash_message(:error, e.message)
